+++ models/Reviews.js
@@ -1,0 +1,1 @@
+const { Sequelize, DataTypes } = require('sequelize');
